% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rob_cloud.R
\name{rob_cloud}
\alias{rob_cloud}
\title{Sentiment Word Cloud}
\usage{
rob_cloud(text)
}
\arguments{
\item{text}{A character vector of text}
}
\value{
A wordcloud
}
\description{
This function creates a word cloud focused around sentiment.
}
\examples{
if (interactive()) {
example_text <- "this bank sucked it was slow and stinky but they gave free cookies so that was good"
rob_cloud(example_text)
}
}
