% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rob_sentiment_score.R
\name{rob_sentiment_score}
\alias{rob_sentiment_score}
\title{Sentiment Score}
\usage{
rob_sentiment_score(text)
}
\arguments{
\item{text}{A character vector of text}
}
\value{
A numeric value representing sentiment
}
\description{
This function returns the sentiment score of your text.
}
\examples{
if (interactive()) {
    example_text <- "this bank sucked! It was slow and stinky!"
    sentiment <- rob_sentiment_score(example_text)
}
}
