% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rob_text.R
\name{rob_text}
\alias{rob_text}
\title{Clean Text}
\usage{
rob_text(text)
}
\arguments{
\item{text}{A character vector of text}
}
\value{
A cleaned character vector of text
}
\description{
This function returns cleaned text.
}
\examples{
example_text <- "This  bank sucked!It2 was slow and stinky!"
clean_text <- rob_text(example_text)
}
