% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rob_token.R
\name{rob_token}
\alias{rob_token}
\title{Tokenize Text}
\usage{
rob_token(text)
}
\arguments{
\item{text}{A character vector of text}
}
\value{
tokens
}
\description{
This function returns tokenized text.
}
\examples{
example_text <- "This  bank sucked!It2 was slow and stinky!"
tokens <- rob_token(example_text)
}
